basic generics support
