+++ .github/workflows/ci.yml
@@ -1,4 +1,4 @@
-# Copyright (c) 2023 Purple Clay
+# Copyright (c) 2023 - 2024 Purple Clay
 #
 # Permission is hereby granted, free of charge, to any person obtaining a copy
 # of this software and associated documentation files (the "Software"), to deal
