+++ combinator.go
@@ -1,5 +1,5 @@
 /*
-Copyright (c) 2023 Purple Clay
+Copyright (c) 2023 - 2024 Purple Clay
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
@@ -27,11 +27,16 @@ import (
 	"strings"
 )
 
+// Result ...
+type Result interface {
+	string | []string
+}
+
 // Combinator is a higher-order function capable of parsing text under a defined
 // condition. Combinators can be combined to form more complex parsers. Upon success,
 // a combinator will return both the unparsed and parsed text. All combinators are
 // strict and must parse its input. Any failure to do so should raise a [CombinatorParseError].
-type Combinator func(string) (string, string, error)
+type Combinator[T Result] func(string) (string, T, error)
 
 const truncateErrAt = 20
 
@@ -67,280 +72,22 @@ func (e CombinatorParseError) Error() string {
 	return buf.String()
 }
 
-// Tag must match a series of characters at the beginning of the input text,
-// in the exact order and case provided.
-//
-//	chomp.Tag("Hello")("Hello, World!")
-//	// (", World!", "Hello", nil)
-func Tag(str string) Combinator {
-	return func(s string) (string, string, error) {
-		if strings.HasPrefix(s, str) {
-			return s[len(str):], str, nil
-		}
+// ParserError defines an error that is raised when a parser
+// fails to parse the input text due to a failed [Combinator]
+type ParserError struct {
+	// Err contains the [CombinatorParseError] that caused the parser to fail.
+	Err error
 
-		return s, "", CombinatorParseError{Input: str, Text: s, Type: "tag"}
-	}
+	// Type of [Parser] that failed.
+	Type string
 }
 
-// Any must match at least one character at the beginning of the input text,
-// from the provided sequence. Parsing immediately stops upon the first
-// unmatched character.
-//
-//	chomp.Any("eH")("Hello, World!")
-//	// ("llo, World!", "He", nil)
-func Any(str string) Combinator {
-	return func(s string) (string, string, error) {
-		pos := 0
-
-	match:
-		for _, sc := range s {
-			for _, strc := range str {
-				if sc == strc {
-					pos = pos + len(string(strc))
-					continue match
-				}
-			}
-
-			break match
-		}
-
-		if pos == 0 {
-			return s, "", CombinatorParseError{Input: str, Text: s, Type: "any"}
-		}
-
-		return s[pos:], s[:pos], nil
-	}
+// Error returns a friendly string representation of the current error.
+func (e ParserError) Error() string {
+	return fmt.Sprintf("%s parser failed. %v", e.Type, e.Err)
 }
 
-// Not must not match at least one character at the beginning of the input
-// text from the provided sequence. Parsing immediately stops upon the
-// first matched character.
-//
-//	chomp.Not("ol")("Hello, World!")
-//	// ("llo, World!", "He", nil)
-func Not(str string) Combinator {
-	return func(s string) (string, string, error) {
-		pos := 0
-
-	match:
-		for _, sc := range s {
-			for _, strc := range str {
-				if sc == strc {
-					break match
-				}
-			}
-
-			pos = pos + len(string(sc))
-		}
-
-		if pos == 0 {
-			return s, "", CombinatorParseError{Input: str, Text: s, Type: "not"}
-		}
-
-		return s[pos:], s[:pos], nil
-	}
-}
-
-// Crlf must match either a CR or CRLF line ending.
-//
-//	chomp.Crlf()("\r\nHello")
-//	// ("Hello", "\r\n", nil)
-func Crlf() Combinator {
-	return func(s string) (string, string, error) {
-		idx := strings.Index(s, "\n")
-		if idx == 0 || (idx == 1 && s[0] == '\r') {
-			return s[idx+1:], s[:idx+1], nil
-		}
-
-		return "", "", CombinatorParseError{Text: s, Type: "crlf"}
-	}
-}
-
-// OneOf must match a single character at the beginning of the text from the
-// provided sequence.
-//
-//	chomp.OneOf("!,eH")("Hello, World!")
-//	// ("ello, World!", "H", nil)
-func OneOf(str string) Combinator {
-	return func(s string) (string, string, error) {
-		if txt := []rune(s); len(txt) > 0 {
-			for _, strc := range str {
-				if txt[0] == strc {
-					pos := len(string(txt[0]))
-					return s[pos:], s[:pos], nil
-				}
-			}
-		}
-
-		return "", "", CombinatorParseError{Input: str, Text: s, Type: "one_of"}
-	}
-}
-
-// NoneOf must not match a single character at the beginning of the text from
-// the provided sequence.
-//
-//	chomp.NoneOf("loWrd!e")("Hello, World!")
-//	// ("ello, World!", "H", nil)
-func NoneOf(str string) Combinator {
-	return func(s string) (string, string, error) {
-		if txt := []rune(s); len(txt) > 0 {
-			for _, strc := range str {
-				if txt[0] == strc {
-					break
-				}
-			}
-
-			pos := len(string(txt[0]))
-			return s[pos:], s[:pos], nil
-		}
-
-		return "", "", CombinatorParseError{Input: str, Text: s, Type: "none_of"}
-	}
-}
-
-// Until will scan the input text for the first occurrence of the provided series
-// of characters. Everything until that point in the text will be matched.
-//
-//	chomp.Until("World")("Hello, World!")
-//	// ("World!", "Hello, ", nil)
-func Until(str string) Combinator {
-	return func(s string) (string, string, error) {
-		if idx := strings.Index(s, str); idx != -1 {
-			return s[idx:], s[:idx], nil
-		}
-
-		return "", "", CombinatorParseError{Input: str, Text: s, Type: "until"}
-	}
-}
-
-// While will scan the input text, testing each character against the provided
-// [Predicate]. Everything until the predicate returns false will be matched.
-//
-//	chomp.While(chomp.IsLetter)("Hello, World!")
-//	// (", World!", "Hello", nil)
-func While(p Predicate) Combinator {
-	return func(s string) (string, string, error) {
-		pos := 0
-		for _, c := range s {
-			if !p(c) {
-				break
-			}
-			pos += len(string(c))
-		}
-
-		if pos == 0 {
-			return "", "", CombinatorParseError{Text: s, Type: "while"}
-		}
-
-		return s[pos:], s[:pos], nil
-	}
-}
-
-// WhileNot will scan the input text, testing each character against the provided
-// [Predicate]. Everything until the predicate returns true will be matched. This
-// is the inverse of [While].
-//
-//	chomp.WhileNot(chomp.IsDigit)("Hello, World!")
-//	// ("", "Hello, World!", nil)
-func WhileNot(p Predicate) Combinator {
-	return func(s string) (string, string, error) {
-		pos := 0
-		for _, c := range s {
-			if p(c) {
-				break
-			}
-			pos += len(string(c))
-		}
-
-		if pos == 0 {
-			return "", "", CombinatorParseError{Text: s, Type: "while_not"}
-		}
-
-		return s[pos:], s[:pos], nil
-	}
-}
-
-// Delimited will match a series of combinators against the input text. The left
-// and right combinators are used to match a delimited sequence and are discarded.
-// Only the text between the delimiters is extracted.
-//
-//	chomp.Delimited(
-//		chomp.Tag("'"),
-//		chomp.Tag("Hello, World!"),
-//		chomp.Tag("'"))("'Hello, World!'")
-//	// ("", "Hello, World!", nil)
-func Delimited(left, str, right Combinator) Combinator {
-	return func(s string) (string, string, error) {
-		rem, _, err := left(s)
-		if err != nil {
-			return s, "", err
-		}
-
-		rem, ext, err := str(rem)
-		if err != nil {
-			return rem, "", err
-		}
-
-		rem, _, err = right(rem)
-		if err != nil {
-			return rem, "", err
-		}
-
-		return rem, ext, nil
-	}
-}
-
-// QuoteDouble will match any text delimited (or surrounded) by a pair
-// of "double quotes". The delimiters are discarded.
-//
-//	chomp.DoubleQuote()(`"Hello, World!"`)
-//	// ("", "Hello, World!", nil)
-func QuoteDouble() Combinator {
-	return func(s string) (string, string, error) {
-		return Delimited(Tag("\""), Until("\""), Tag("\""))(s)
-	}
-}
-
-// QuoteSingle will match any text delimited (or surrounded) by a pair
-// of 'single quotes'. The delimiters are discarded.
-//
-//	chomp.QuoteSingle()("'Hello, World!'")
-//	// ("", "Hello, World!", nil)
-func QuoteSingle() Combinator {
-	return func(s string) (string, string, error) {
-		return Delimited(Tag("'"), Until("'"), Tag("'"))(s)
-	}
-}
-
-// BracketSquare will match any text delimited (or surrounded) by a pair
-// of [square brackets]. The delimiters are discarded.
-//
-//	chomp.BracketSquare()("[Hello, World!]")
-//	// ("", "Hello, World!", nil)
-func BracketSquare() Combinator {
-	return func(s string) (string, string, error) {
-		return Delimited(Tag("["), Until("]"), Tag("]"))(s)
-	}
-}
-
-// Parentheses will match any text delimited (or surrounded) by a pair
-// of (parentheses). The delimiters are discarded.
-//
-//	chomp.Parentheses()("(Hello, World!)")
-//	// ("", "Hello, World!", nil)
-func Parentheses() Combinator {
-	return func(s string) (string, string, error) {
-		return Delimited(Tag("("), Until(")"), Tag(")"))(s)
-	}
-}
-
-// BracketAngled will match any text delimited (or surrounded) by a pair
-// of <angled brackets>. The delimiters are discarded.
-//
-//	chomp.BracketAngled()("<Hello, World!>")
-//	// ("", "Hello, World!", nil)
-func BracketAngled() Combinator {
-	return func(s string) (string, string, error) {
-		return Delimited(Tag("<"), Until(">"), Tag(">"))(s)
-	}
+// Unwrap returns the inner [CombinatorParseError].
+func (e ParserError) Unwrap() error {
+	return e.Err
 }
